Fixed a bunch of compile warnings and general clean up.

+++ SC_backend/slow_control_code/ADAM6000/ADAM6017.c
@@ -5,13 +5,6 @@
 /* Copyright 2006, 2007, 2009 */
 /* James public licence. */
 
-#include <stdlib.h>
-#include <stdio.h>
-#include <sys/fcntl.h>
-#include <unistd.h>
-#include <time.h>
-#include <signal.h>
-
 #include "SC_db_interface.h"
 #include "SC_aux_fns.h"
 #include "SC_sensor_interface.h"
@@ -35,38 +28,38 @@ const uint16_t num_DO = 2;
 #define _def_read_sensor
 int read_sensor(struct inst_struct *i_s, struct sensor_struct *s_s, double *val_out)
 {  
-    int ret;
+  int ret;
 
-    if (strncmp(s_s->subtype, "ADC", 3) == 0)           // Analogue in     
-	return(read_ADC(s_s, ADC_start_address,  num_ADC, val_out));
+  if (strncmp(s_s->subtype, "ADC", 3) == 0)           // Analogue in     
+    return(read_ADC(s_s, ADC_start_address,  num_ADC, val_out));
     
-    else if (strncmp(s_s->subtype, "LLADC", 3) == 0)  
-      {      
-	ret = read_ADC(s_s, ADC_start_address,  num_ADC, val_out);
-	*val_out = pow(10, (double)(*val_out) -5.0);
-	return(ret);
-      }
+  else if (strncmp(s_s->subtype, "LLADC", 3) == 0)  
+    {      
+      ret = read_ADC(s_s, ADC_start_address,  num_ADC, val_out);
+      *val_out = pow(10, *val_out - 5.0);
+      return(ret);
+    }
 
-    else if (strncmp(s_s->subtype, "DO", 2) == 0)       // Digital Out
-	return(read_DO(s_s, DO_start_address,  num_DO, val_out));
+  else if (strncmp(s_s->subtype, "DO", 2) == 0)       // Digital Out
+    return(read_DO(s_s, DO_start_address,  num_DO, val_out));
 
-    else
+  else
     {
-	fprintf(stderr, "Wrong subtype for %s \n", s_s->name);
-	return(1);
+      fprintf(stderr, "Wrong subtype for %s \n", s_s->name);
+      return(1);
     }
 }
 
 #define _def_set_sensor
 int set_sensor(struct inst_struct *i_s, struct sensor_struct *s_s)
 {
-    if (strncmp(s_s->subtype, "DO", 2) == 0)       // Digital out 
-	return(write_DO(s_s, DO_start_address,  num_DO));
+  if (strncmp(s_s->subtype, "DO", 2) == 0)       // Digital out 
+    return(write_DO(s_s, DO_start_address,  num_DO));
 
-    else
+  else
     {
-	fprintf(stderr, "Wrong subtype for %s \n", s_s->name);
-	return(1);
+      fprintf(stderr, "Wrong subtype for %s \n", s_s->name);
+      return(1);
     }
 }
